Working now

+++ src/dunzo/RecipeIngredient.java
@@ -2,15 +2,15 @@ package dunzo;
 
 public class RecipeIngredient {
 	private Ingredient ingredient;
-	private int quantity;
-	RecipeIngredient(Ingredient ingredient, int quantity){
+	private long quantity;
+	RecipeIngredient(Ingredient ingredient, long quantity){
 		this.ingredient = ingredient;
 		this.quantity = quantity;
 	}
 	public Ingredient getIngredient() {
 		return this.ingredient;
 	}
-	public int getQuantity() {
+	public long getQuantity() {
 		return this.quantity;
 	}
 }
